+++ torchio/__init__.py
@@ -2,11 +2,12 @@
 
 __author__ = """Fernando Perez-Garcia"""
 __email__ = 'fernando.perezgarcia.17@ucl.ac.uk'
-__version__ = '0.2.0'
+__version__ = '0.6.1'
 
 from . import utils
 from . import sampler
 from . import inference
+from .torchio import *
 from .queue import Queue
 from .transforms import *
 from .dataset import ImagesDataset