+++ torchio/transforms/__init__.py
@@ -2,10 +2,12 @@
 from .random_flip import RandomFlip
 from .random_noise import RandomNoise
 from .random_affine import RandomAffine
-from .random_elastic_deform import RandomElasticDeformation
+from .random_motion import RandomMotion
+from .random_bias_field import RandomBiasField
+from .random_elastic_deformation import RandomElasticDeformation
 from .interpolation import Interpolation
 
 # Intensity transforms
 from .rescale import Rescale
 from .z_normalization import ZNormalization
-from .histogram_standardisation import HistogramStandardisation
+from .histogram_standardization import HistogramStandardization